testing sol 2level

+++ hw3/hw3.cpp
@@ -46,7 +46,7 @@ struct entry_2_bit
 {
     bool prediction;
     char state;
-    UINT64 ht;
+    UINT16 ht;
 } BPB_2_bit[SIZE];
 
 /* initialize the BPB, not taken by default*/
